Rename folder Configs to Common from Integration tests
