+++ tests/N-Tier.Api.IntegrationTests/Tests/TodoListEndpointTests.cs
@@ -6,8 +6,9 @@ using System.Threading.Tasks;
 using FizzWare.NBuilder;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using N_Tier.Api.IntegrationTests.Config;
-using N_Tier.Api.IntegrationTests.Config.Constants;
+using N_Tier.API;
+using N_Tier.Api.IntegrationTests.Common;
+using N_Tier.Api.IntegrationTests.Common.Constants;
 using N_Tier.Api.IntegrationTests.Helpers;
 using N_Tier.Application.Models;
 using N_Tier.Application.Models.TodoList;
