+++ tests/N-Tier.Api.IntegrationTests/Tests/UsersEndpointTests.cs
@@ -5,8 +5,9 @@ using FizzWare.NBuilder;
 using FluentAssertions;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
-using N_Tier.Api.IntegrationTests.Config;
-using N_Tier.Api.IntegrationTests.Config.Constants;
+using N_Tier.API;
+using N_Tier.Api.IntegrationTests.Common;
+using N_Tier.Api.IntegrationTests.Common.Constants;
 using N_Tier.Api.IntegrationTests.Helpers;
 using N_Tier.Application.Models;
 using N_Tier.Application.Models.User;
